Refactor client to not be global varaible
